Beta Version 0.1.1.4
Added settings page, to look for arma3.exe file path in case you have a
different installation path. This should also allow the user to launch
the game correctly from the correct ArmA 3 installation path.

It saves the settings page and if you want the launcher to close or stay
open when launching the game.

Also added the Mods check list box - This List will display any mods
within the Arma 3 Directory. It looks for any folder that starts with
the @ in it's name.